requirements.txt
praw    #Python Reddit API Wrapper for fetching user posts and comments
python-dotenv   #Environment Management: to securely load API keys
google-generativeai #LLM: Google Gemini 2.0 Flash Lite - used to generate a detailed user persona from Reddit content
